requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.2.1
email-validator==2.1.0
werkzeug==2.3.7
python-dotenv==1.0.0
